crosier-chrome: Adding tags to Crosier test metadata and other fixes.
Tags are required for mapping Crosier tests into CI/CQ suites.

Also fixed:
- Added metadata for a few more HW agnostic tests to start integration.
- Fixed init script to check if initialization is needed.

BUG=b:294097177
TEST=local build and test

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5143786
Commit-Queue: Alex Bergman <abergman@google.com>
Tested-by: Alex Bergman <abergman@google.com>

+++ crosier-chrome/init_env.sh
@@ -4,7 +4,16 @@
 # found in the LICENSE file.
 
 # Script to initialize Crosier environment.
-# This script should run on device/VM.
+# This script runs on device/VM.
+#
+# The directory structure after initialization should be:
+#   /usr/local/libexec/crosier/... - Crosier testing directory
+#     -> chromeos_integration_tests - test binary
+#     -> test_sudo_helper* - sudo helper script
+#     -> symlink to all Chrome libs from /opt/google/chrome/*
+#     -> symlink to libtest_trace_processor.so
+#     -> test_fonts
+#   /usr/local/chrome/test/data/chromeos/web_handwriting - handwriting libs
 
 set -e
 
@@ -12,12 +21,24 @@ set -e
 HERE="$(dirname "$(realpath -e "${BASH_SOURCE[0]}")")"
 cd "${HERE}"
 
-# Link and copy chrome libs and other files to correct locations
-ln -s /opt/google/chrome/* .
-ln -s /usr/local/libexec/chrome-binary-tests/libtest_trace_processor.so .
-mkdir -p ../../chrome/test/data/chromeos
-mv web_handwriting ../../chrome/test/data/chromeos/
+# Uncompress the test binary if it's still compressed.
+COMPRESSED_NAME="chromeos_integration_tests.tar.bz2"
+if [[ -e "${COMPRESSED_NAME}" ]]; then
+  tar xfj "${COMPRESSED_NAME}"
+  rm "${COMPRESSED_NAME}"
+fi
 
-# Set permissions and folder owners
+# Link and copy chrome libs and other files to correct locations.
+if [[ ! -L "./chrome" ]]; then
+  ln -s /opt/google/chrome/* .
+  ln -s /usr/local/libexec/chrome-binary-tests/libtest_trace_processor.so .
+fi
+
+if [[ -d ./web_handwriting ]]; then
+  mkdir -p ../../chrome/test/data/chromeos
+  mv web_handwriting ../../chrome/test/data/chromeos/
+fi
+
+# Set permissions and folder owners.
 chown -R chronos: ../../chrome
 chown -R chronos: ..
